requirements.txt
openai
whisper
torch
torchaudio
python-dotenv
